Fixed guide on how to enable unattended-upgrades on Raspberry Pi OS
Existing debconf configuration for packages like unattended-upgrades cannot
be changed using dpkg-reconfigure with noninteractive frontend. "It is by
design of debconf that settings on the system take precedence over any values
set in the debconf database. There is a valid use case for being able to
preseed the set of modules that you want to install, but it is difficult to
implement this while maintaining the requirement to respect any local changes
to the config files." [1] Refer to debconf.sh for rationale.

Ref.: https://bugs.launchpad.net/ubuntu/+source/pam/+bug/682662/comments/1